+++ pytsite/formatters/__init__.py
@@ -1,6 +1,6 @@
 """PytSite Formatter
 """
-from ._formatters import Formatter, Bool, Int, PositiveInt, Float, PositiveFloat, Str, DateTime, JSON, \
+from ._formatters import Formatter, Bool, Int, PositiveInt, AboveZeroInt, Float, PositiveFloat, Str, DateTime, JSON, \
     JSONArrayToList, JSONArrayToTuple, JSONObjectToDict
 
 __author__ = 'Alexander Shepetko'